parser finally works!
After 26 long hours of trying.

+++ Test/hej/TestFileInterpreter.java
@@ -59,9 +59,10 @@ public class TestFileInterpreter {
             }""";
         ArrayList<String> objectsFound = FileInterpreter.findObjects(first, "UnitTest testParse of FileInterpreter", false);
         Object hashMap = FileInterpreter.parse(objectsFound.get(0), "UnitTest testParse of FileInterpreter");
+
         Assertions.assertEquals("small rat", HashMapExplorer.getString(hashMap, "name"));
-        Assertions.assertEquals(2, HashMapExplorer.getNumber(hashMap, "defence"));
-        Assertions.assertEquals(2, HashMapExplorer.getNumber(hashMap,"damage.min"));
+        Assertions.assertEquals(2, HashMapExplorer.getNumber(hashMap, "defence").intValue());
+        Assertions.assertEquals(2.0, HashMapExplorer.getNumber(hashMap,"damage.min").floatValue());
     }
 
     @Test
